pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
nfc = "nfc:nutfeex_cli"

[tool.bandit]
exclude_dirs = ["tests", ".tox"]

[tool.black]
line-length = 79

[tool.coverage.run]
branch = true
source = "."
omit = ".tox"

[tool.flake8]
statistics = true
exclude = [
    "__pycache__",
    ".git",
    ".tox",
]

[tool.coverage.report]
show_missing = true
precision = 3
fail_under = 100

[tool.poetry]
name = "nfc"
version = "0.1.0"
description = "Nutfeex CLI"
authors = ["Luis Muñoz <luis.munoz.villarreal@gmail.com>"]
packages = [
    { include = "nfc" },
]


[tool.poetry.dependencies]
click = "^8.1.7"
configobj = "^5.0.8"
docker = "^7.0.0"
dockerpty = "^0.4.1"
gitpython = "^3.1.41"
pydantic = "^2.6.1"
python = "^3.11"
sh = "^2.0.6"

[tool.poetry.group.dev.dependencies]
bandit = {extras = ["toml"], version = "^1.7.7"}
black = "^24.2.0"
flake8 = "^7.0.0"
flake8-isort = "^6.1.1"
flake8-pyproject = "^1.2.3"
ipython = "^8.21.0"
ipdb = "^0.13.13"
mypy = "^1.8.0"
pydocstyle = "^6.3.0"
pylint = "^3.0.3"
pytest-cov = "^4.1.0"
pytest-factoryboy = "^2.6.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
pytest = "^8.0.0"
tox = "^4.12.1"

[tool.pydocstyle]
convention = "google"
match = ".*\\.py"
add-ignore = "D107,D104,D106"

[tool.pylint.MASTER]
load-plugins = "pylint.extensions.docparams"
accept-no-param-doc = "no"
accept-no-raise-doc = "no"
accept-no-return-doc = "no"
jobs = 0
reports = "y"
max-line-length = 80

[tool.pylint."MESSAGES CONTROL"]
disable = "model-no-explicit-unicode"


[tool.pytest.ini_options]
addopts = [
    "--cov",
    "--cov-report term",
    "--no-cov-on-fail",
    "--strict-markers",
    "-Werror",
]
